+++ modules/constants.js
@@ -49,70 +49,73 @@ export const SPELL_BETTER = {
         custom : []
     },
     standardCategories : { 
+        //.prop is needed for now for display
+//FIXME: For now, we are completely overriding the custom categories every time you restart        
         spell0: { label: "DND5E.SpellLevel0", canCreate: true, canPrepare: false, 
-                dataset: {type: "spell", level: 0},
+                templateItemData: {type: "spell", level: 0},
                 filterSets: [
                     {filterSet : "levels", filters: ["Cantrip"]}
                 ],
                 order: 0, prop: "spell0", slots: "&infin;", uses: "&infin;", usesSlots: false, type: "spell",  isCollapsed: true },
         spell1: { label: "SPELL_BETTER.SpellLevel1", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 1},
+                templateItemData: {type: "spell", level: 1},
                 filterSets: [
                     {filterSet : "levels", filters: ["1st Level"]}
                 ],
+//FIXME: Slots, canCreate etc etc is all hard-coded right now, but somehow we need to use the actual values from the sections                
                 order: 10, prop: "spell1", slots: "4", uses: "4", usesSlots: true, type: "spell",  isCollapsed: false },     
         spell2: { label: "SPELL_BETTER.SpellLevel2", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 2},
+                templateItemData: {type: "spell", level: 2},
                 filterSets: [
                     {filterSet : "levels", filters: ["2nd Level"]}
                 ],                
                 order: 20, prop: "spell2", slots: "3", uses: "3", usesSlots: true, type: "spell",  isCollapsed: false },
         spell3: { label: "SPELL_BETTER.SpellLevel3", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 3},
+                templateItemData: {type: "spell", level: 3},
                 filterSets: [
                     {filterSet : "levels", filters: ["3rd Level"]}
                 ],                
                 order: 30, prop: "spell3", slots: "3", uses: "3", usesSlots: true, type: "spell",  isCollapsed: false },
         spell4: { label: "SPELL_BETTER.SpellLevel4", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 4},
+                templateItemData: {type: "spell", level: 4},
                 filterSets: [
                     {filterSet : "levels", filters: ["4th Level"]}
                 ],                
                 order: 40, prop: "spell4", slots: "0", uses: "0", usesSlots: true, type: "spell",  isCollapsed: false },
         spell5: { label: "SPELL_BETTER.SpellLevel5", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 5},
+                templateItemData: {type: "spell", level: 5},
                 filterSets: [
                     {filterSet : "levels", filters: ["5th Level"]}
                 ],                
                 order: 50, prop: "spell5", slots: "0", uses: "0", usesSlots: true, type: "spell",  isCollapsed: false },
         spell6: { label: "SPELL_BETTER.SpellLevel6", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 6},
+                templateItemData: {type: "spell", level: 6},
                 filterSets: [
                     {filterSet : "levels", filters: ["6th Level"]}
                 ],                
                 order: 60, prop: "spell6", slots: "0", uses: "0", usesSlots: true, type: "spell",  isCollapsed: false },
         spell7: { label: "SPELL_BETTER.SpellLevel7", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 7},
+                templateItemData: {type: "spell", level: 7},
                 filterSets: [
                     {filterSet : "levels", filters: ["7th Level"]}
                 ],                
                 order: 70, prop: "spell7", slots: "0", uses: "0", usesSlots: true, type: "spell",  isCollapsed: false },
         spell8: { label: "SPELL_BETTER.SpellLevel8", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 8},
+                templateItemData: {type: "spell", level: 8},
                 filterSets: [
                     {filterSet : "levels", filters: ["8th Level"]}
                 ],                
                 order: 80, prop: "spell8", slots: "0", uses: "0", usesSlots: true, type: "spell",  isCollapsed: false },  
         spell9: { label: "SPELL_BETTER.SpellLevel9", canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 9},
+                templateItemData: {type: "spell", level: 9},
                 filterSets: [
                     {filterSet : "levels", filters: ["9th Level"]}
                 ],                
                 order: 90, prop: "spell8", slots: "0", uses: "0", usesSlots: true, type: "spell",  isCollapsed: false },                                                                                                                                          
         wanted: { label: "SPELL_BETTER.Category.Wanted",canCreate: true, canPrepare: true, 
-                dataset: {type: "spell", level: 0},
+                templateItemData: {type: "spell", level: 0},
                 filterSets: [
-                    {filterSet : "custom", filters: ["wanted"]}
+                    {filterSet : "levels", filters: ["4th Level","5th Level","6th Level","7th Level","8th Level","9th Level"]}
                 ],                
                 order: 1000, prop: "spell1", slots: "&infin;", uses: "&infin;", usesSlots: false, type: "spell",  isCollapsed: true }
     }