+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # RELEASE NOTES
 ## 0.5.1 21-Dec-2020
-- Incorporated Inventory+ into Spell Better (converted to aply to spells)
+- Incorporated Inventory+ into Spell Better (converted to apply to spells)
 ## 0.5.0 
 - Forked from ogl-5e-character-sheet???